added the ablity to play sound
added a audio manager class that manages the sound effects and the
background music.

added a folder for sound effects and a open source sound effect of an
explosion.

changed the way that keyboard inputs are handled so that move is
processed every time the loop resets and the rest of the buttons are
when when pressed down.