added an animation manager
the animation manager will update the differnt objects possitions as
they follow diffent routes. It will also handle the plannin and changing
of routes that differnt objects will follow.

This can handle the rendering objects and soon the sound objects

the path finding in the city should also be passed to this since this
class will be needing to deal with it.

still needs the code to updated an object based on a define routine
and proper integration into the city class