added new type of generation (building)
The building generator takes a block of land and will construct walls
and spot for door. The walls can be smaller than the full plot of land
given to it. This will eventally replace the other functions that
generate walls around the roads

added new model for section of wall that has a spot for a door in it