added all the objects to the engine
added the loading dock to the side of the building that the dock spawns
on and the loading dock floor will also spawn there.

- the model is coming in backwards in some possitions

added loading dock doors to the engine and added a special door
routine/actor to allow the doors to slide open.

-currently only spawns on the dock_side = 1

added a wall clock to the corner of the building